centralized validate_spec_for_new_user_ns
