+++ crates/libcontainer/src/container/init_builder.rs
@@ -212,8 +212,6 @@ impl InitContainerBuilder {
             validate_idmapped_mounts(mounts, spec.linux().as_ref(), &*syscall)?;
         }
 
-        utils::validate_spec_for_new_user_ns(spec, &*syscall)?;
-
         Ok(())
     }
